Refactored the requestToursForYear method. The tours for 2015 are now correctly displayed automatically (either read from a file or requested from phish.in) when the tour picker is displayed.
What I need to do though, is to only make the request or read the file if the tour picker is displayed at all. Currently, because the tour picker is part of the view, the requestToursForYear method is called when the view is loaded, not when the tour picker is displayed. So, a request or file read, will be made, regardless of whether the tour picker is actually displayed. This should be fixed.