The setlist VC is working. It's not pretty, but I can continue to work on it.
If possible, I'll now implement a segue to a VC where you can see all the shows, grouped by tour, where a specific song was played. So, you see a setlist, you tap on a song, you see the song history VC, and you can then tap on a particular version in the history, and be taken back to the map, where now the map is displaying that tour, in which the show where the version of that song was played, is a part of.